Dynamic creation of cars and routes.
This is one of the first tryouts that was successful in the creation of cars while the simulation is running. TRACI for Python is used.

A "Scheduler" like implementation has been integrated for executing different actions at several times.

On the 20 steps task there is a way to get access to all the current dynamic vehicles.

Original netfile was changed to add a new double lane road. Code  was changed accordingly.

All double quotes have been changed to single ones according to flake8 rules.