Rerouting has been implemented on all vehicle types.
Rerouting was enabled by default on all vehicles by modifying the *.sumocfg file. It is also possible to create a new instance of a vehicle and enable the rerouting device from a genericParameterCall from TRACI (see runner.py)

Also, every 100 step cycles the first DynamicVeh will be redirectioned to the start edge, just for the matter of testing out the functions.